Add tabs to UserProfileEdit page, extract UserProfile form to separate component
