+++ src/modules/users/components/UserProfileEdit/UserProfileEdit.tsx
@@ -1,11 +1,13 @@
 import React, { useCallback } from 'react';
-import { defineMessages } from 'react-intl';
+import { defineMessages, FormattedMessage } from 'react-intl';
 import * as yup from 'yup';
 import { Redirect } from 'react-router-dom';
 
 import CopyableAddress from '~core/CopyableAddress';
 import UserMention from '~core/UserMention';
 import Heading from '~core/Heading';
+import { Tab, TabList, TabPanel, Tabs } from '~core/Tabs';
+
 import {
   FieldSet,
   Form,
@@ -21,63 +23,24 @@ import { LANDING_PAGE_ROUTE } from '~routes/index';
 
 import UserProfileSpinner from '../UserProfile/UserProfileSpinner';
 import Sidebar from './Sidebar';
+import UserMainSettings from './UserMainSettings';
 import styles from './UserProfileEdit.css';
 
 const MSG = defineMessages({
-  heading: {
-    id: 'users.UserProfileEdit.heading',
-    defaultMessage: 'Profile',
+  headingMain: {
+    id: 'users.UserProfileEdit.headingMain',
+    defaultMessage: 'User profile',
   },
-  labelWallet: {
-    id: 'users.UserProfileEdit.labelWallet',
-    defaultMessage: 'Your Wallet',
-  },
-  labelUsername: {
-    id: 'users.UserProfileEdit.labelUsername',
-    defaultMessage: 'Unique Username',
-  },
-  labelName: {
-    id: 'users.UserProfileEdit.labelName',
-    defaultMessage: 'Name',
-  },
-  labelBio: {
-    id: 'users.UserProfileEdit.labelBio',
-    defaultMessage: 'Bio',
-  },
-  labelWebsite: {
-    id: 'users.UserProfileEdit.labelWebsite',
-    defaultMessage: 'Website',
-  },
-  labelLocation: {
-    id: 'users.UserProfileEdit.labelLocation',
-    defaultMessage: 'Location',
+  headingAdvance: {
+    id: 'users.UserProfileEdit.headingAdvance',
+    defaultMessage: 'Advanced settings',
   },
 });
 
 const displayName = 'users.UserProfileEdit';
 
-interface FormValues {
-  displayName?: string;
-  bio?: string;
-  website?: string;
-  location?: string;
-}
-
-const validationSchema = yup.object({
-  bio: yup.string().nullable(),
-  displayName: yup.string().nullable(),
-  location: yup.string().nullable(),
-  website: yup.string().url().nullable(),
-});
-
 const UserProfileEdit = () => {
   const { walletAddress, ethereal } = useLoggedInUser();
-
-  const [editUser] = useEditUserMutation();
-  const onSubmit = useCallback(
-    (profile: FormValues) => editUser({ variables: { input: profile } }),
-    [editUser],
-  );
 
   const user = useUser(walletAddress);
 
@@ -94,72 +57,26 @@ const UserProfileEdit = () => {
       appearance={{ theme: 'alt' }}
       asideContent={<Sidebar user={user} />}
     >
-      <Heading
-        appearance={{ theme: 'dark', size: 'medium' }}
-        text={MSG.heading}
-      />
-      <Form<FormValues>
-        initialValues={{
-          displayName: user.profile.displayName || undefined,
-          bio: user.profile.bio || undefined,
-          website: user.profile.website || undefined,
-          location: user.profile.location || undefined,
-        }}
-        onSubmit={onSubmit}
-        validationSchema={validationSchema}
-      >
-        {({ status, isSubmitting }) => (
-          <div className={styles.main}>
-            <FieldSet>
-              <InputLabel label={MSG.labelWallet} />
-              <CopyableAddress appearance={{ theme: 'big' }} full>
-                {user.profile.walletAddress}
-              </CopyableAddress>
-            </FieldSet>
-            <div className={styles.usernameContainer}>
-              <InputLabel label={MSG.labelUsername} />
-              <UserMention
-                username={user.profile.username || user.profile.walletAddress}
-                title={user.profile.username || user.profile.walletAddress}
-                hasLink={false}
-                data-test="userProfileUsername"
-              />
-            </div>
-            <FieldSet className={styles.inputFieldSet}>
-              <Input
-                label={MSG.labelName}
-                name="displayName"
-                data-test="userSettingsName"
-              />
-              <Textarea
-                label={MSG.labelBio}
-                name="bio"
-                maxLength={160}
-                data-test="userSettingsBio"
-              />
-              <Input
-                label={MSG.labelWebsite}
-                name="website"
-                data-test="userSettingsWebsite"
-              />
-              <Input
-                label={MSG.labelLocation}
-                name="location"
-                data-test="userSettingsLocation"
-              />
-            </FieldSet>
-            <FieldSet>
-              <Button
-                type="submit"
-                text={{ id: 'button.save' }}
-                loading={isSubmitting}
-                data-test="userSettingsSubmit"
-              />
-            </FieldSet>
-            <FormStatus status={status} />
-          </div>
-        )}
-      </Form>
+      <Tabs>
+        <TabList >
+          <Tab>
+            <FormattedMessage
+              {...MSG.headingMain}
+            />
+          </Tab>
+          <Tab>
+            <FormattedMessage
+              {...MSG.headingAdvance}
+            />
+          </Tab>
+        </TabList>
+        <TabPanel>
+          <UserMainSettings user={user} />
+        </TabPanel>
+        <TabPanel>
+          <h2>Any content 2</h2>
+        </TabPanel>
+      </Tabs>
     </ProfileTemplate>
   );
 };
